--- module/VuFind/src/VuFind/Db/Table/AccessToken.php
@@ -1,142 +1,0 @@
-<?php
-
-/**
- * Table Definition for access_token
- *
- * PHP version 8
- *
- * Copyright (C) Villanova University 2010.
- * Copyright (C) The National Library of Finland 2022.
- *
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License version 2,
- * as published by the Free Software Foundation.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
- *
- * @category VuFind
- * @package  Db_Table
- * @author   Demian Katz <demian.katz@villanova.edu>
- * @author   Ere Maijala <ere.maijala@helsinki.fi>
- * @license  http://opensource.org/licenses/gpl-2.0.php GNU General Public License
- * @link     https://vufind.org Main Page
- */
-
-namespace VuFind\Db\Table;
-
-use Laminas\Db\Adapter\Adapter;
-use VuFind\Db\Row\AccessToken as AccessTokenRow;
-use VuFind\Db\Row\RowGateway;
-
-/**
- * Table Definition for access_token
- *
- * @category VuFind
- * @package  Db_Table
- * @author   Demian Katz <demian.katz@villanova.edu>
- * @author   Ere Maijala <ere.maijala@helsinki.fi>
- * @license  http://opensource.org/licenses/gpl-2.0.php GNU General Public License
- * @link     https://vufind.org Main Site
- */
-class AccessToken extends Gateway
-{
-    use ExpirationTrait;
-
-    /**
-     * Constructor
-     *
-     * @param Adapter       $adapter Database adapter
-     * @param PluginManager $tm      Table manager
-     * @param array         $cfg     Laminas configuration
-     * @param RowGateway    $rowObj  Row prototype object (null for default)
-     * @param string        $table   Name of database table to interface with
-     */
-    public function __construct(
-        Adapter $adapter,
-        PluginManager $tm,
-        $cfg,
-        ?RowGateway $rowObj = null,
-        $table = 'access_token'
-    ) {
-        parent::__construct($adapter, $tm, $cfg, $rowObj, $table);
-    }
-
-    /**
-     * Retrieve an object from the database based on id and type; create a new
-     * row if no existing match is found.
-     *
-     * @param string $id     Token ID
-     * @param string $type   Token type
-     * @param bool   $create Should we create rows that don't already exist?
-     *
-     * @return ?AccessTokenRow
-     */
-    public function getByIdAndType(
-        string $id,
-        string $type,
-        bool $create = true
-    ): ?AccessTokenRow {
-        $row = $this->select(['id' => $id, 'type' => $type])->current();
-        if ($create && empty($row)) {
-            $row = $this->createRow();
-            $row->id = $id;
-            $row->type = $type;
-            $row->created = date('Y-m-d H:i:s');
-        }
-        return $row;
-    }
-
-    /**
-     * Add or replace an OpenID nonce for a user
-     *
-     * @param int     $userId User ID
-     * @param ?string $nonce  Nonce
-     *
-     * @return void
-     */
-    public function storeNonce(int $userId, ?string $nonce)
-    {
-        $row = $this->getByIdAndType($userId, 'openid_nonce');
-        $row->created = date('Y-m-d H:i:s');
-        $row->user_id = $userId;
-        $row->data = json_encode(compact('nonce'));
-        $row->save();
-    }
-
-    /**
-     * Retrieve an OpenID nonce for a user
-     *
-     * @param int $userId User ID
-     *
-     * @return ?string
-     */
-    public function getNonce(int $userId): ?string
-    {
-        if ($row = $this->getByIdAndType($userId, 'openid_nonce', false)) {
-            $data = json_decode($row->data, true);
-            return $data['nonce'] ?? null;
-        }
-        return null;
-    }
-
-    /**
-     * Update the select statement to find records to delete.
-     *
-     * @param Select $select    Select clause
-     * @param string $dateLimit Date threshold of an "expired" record in format
-     * 'Y-m-d H:i:s'.
-     *
-     * @return void
-     */
-    protected function expirationCallback($select, $dateLimit)
-    {
-        $select->where->lessThan('created', $dateLimit);
-    }
-}
